delete sign in and sign up
